Restarting with a new template
Restarting with a new library management system template, as having very less time, as well as it is having almost all the required elements and pages, just need some modifications. here is the original and starting version of the same, need to modify as needed and connect to the database with backend (php), asap.